requirements.txt
discord.py
opencc-python-reimplemented
Pillow
